pyproject.toml
[tool.poetry]
name = "rag"
version = "0.1.0"
description = ""
authors = ["Gustaf Rydholm <gustaf.rydholm@gmail.com>"]
readme = "README.md"
packages = [{include = "rag"}]

[tool.poetry.dependencies]
python = "^3.11"
ollama = "^0.1.8"
langchain-community = "^0.0.31"
langchain = "^0.1.14"
jupyterlab-vim = "^4.1.3"
pypdf = "^4.1.0"
numpy = "^1.26.4"
psycopg = "^3.1.18"
python-dotenv = "^1.0.1"
qdrant-client = "^1.8.2"
loguru = "^0.7.2"
streamlit = "^1.33.0"
cohere = "^5.2.3"
tqdm = "^4.66.2"
click = "^8.1.7"
sentence-transformers = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
